Move bounding method functionality to solvers/bounder.xpp
