+++ source/solvers/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_library(ariadne-solvers OBJECT
     solver.cpp
     integrator.cpp
+    bounder.cpp
     constraint_solver.cpp
     simplex_algorithm.cpp
     linear_programming.cpp
